Optimizado

+++ assets/js/index.js
@@ -1,4 +1,4 @@
-//Entrega Final 2
+//Incorporando Librerias
 
 let contenedor_productos = document.getElementById("productos")
 let contenedor_carrito = document.getElementById("carrito")
@@ -12,13 +12,7 @@ cargoProductos();
 
 //Cargo los productos del localStorage
 function cargarCarrito() {
-    if (localStorage.getItem("StorageProductos") !== null) {
-        carrito = JSON.parse(localStorage.getItem("StorageProductos"));
-        return;
-    } else {
-        localStorage.setItem("StorageFavoritos", JSON.stringify(carrito));
-        return;
-    }
+    carrito = JSON.parse(localStorage.getItem("StorageProductos")) || localStorage.setItem("StorageProductos", JSON.stringify(carrito));
 }
 
 //Cargo los productos
@@ -33,34 +27,71 @@ function cargoProductos() {
                 <p>${productos.name}</p>
                 <p>$ ${productos.precio}</P>
                 <div class="flex-row">
+                    <input type="numero" id="cant${productos.id}" placeholder="Cantididad..."> 
                     <button class="btn-comprar" onClick="seleccionProductos(${productos.id})">Comprar</button>
                 </div>
             </div>
             
         </div>
-            `
+        `
     });
 }
 
 //Agrego los productos seleccionados al carrito
 
 function seleccionProductos(id_seleccionado) {
+    (document.getElementById("cant1")) ? cant1 = document.getElementById("cant1").value: cant1 = 0;
+    (document.getElementById("cant2")) ? cant2 = document.getElementById("cant2").value: cant2 = 0;
+    (document.getElementById("cant3")) ? cant3 = document.getElementById("cant3").value: cant3 = 0;
+    (document.getElementById("cant4")) ? cant4 = document.getElementById("cant4").value: cant4 = 0;
+    (document.getElementById("cant5")) ? cant5 = document.getElementById("cant5").value: cant5 = 0;
+    (document.getElementById("cant6")) ? cant6 = document.getElementById("cant6").value: cant6 = 0;
+
+    cant1 > 0 ? cantidad = +(cant1) : false
+    cant2 > 0 ? cantidad = +(cant2) : false
+    cant3 > 0 ? cantidad = +(cant3) : false
+    cant4 > 0 ? cantidad = +(cant4) : false
+    cant5 > 0 ? cantidad = +(cant5) : false
+    cant6 > 0 ? cantidad = +(cant6) : false
+
     let indice = id_seleccionado - 1;
 
     let objeto_seleccionado = {};
-    objeto_seleccionado = productos[indice];
-    console.log(objeto_seleccionado);
+    objeto_seleccionado = {
+        ...productos[indice],
+        cantidad: cantidad
+    }
+    console.log(objeto_seleccionado)
+    if (!carrito.some((el) => el.id == id_seleccionado)) { //Sino existe el producto en el carrito lo agrego
 
-    carrito.push(objeto_seleccionado);
-    console.log(carrito)
+        carrito.push(objeto_seleccionado);
+        console.log(carrito)
+        localStorage.setItem("StorageProductos", JSON.stringify(carrito));
 
-    localStorage.setItem("StorageProductos", JSON.stringify(carrito));
-    location.reload()
+        Toastify({ //Libreria toastify
+            text: "Producto agregado al carrito",
+            duration: 2000,
+            gravity: "top", // `top` or `bottom`
+            position: "right", // `left`, `center` or `right`
+            className: "notificacion",
+        }).showToast();
 
+        setTimeout(() => {
+            location.reload()
+        }, 1000)
+
+    } else { //Si existe el producto en el carrito
+        Swal.fire({ //Libreria sweetalert
+            text: 'Debes eliminar el producto del carrito primero',
+            icon: 'info',
+            confirmButtonText: 'ok'
+        })
+    }
 }
 
 //Muestro los productos del carrito
 function mostrarCarrito() {
+
     carrito.forEach((carrito) => {
         contenedor_carrito.innerHTML += `
         <div class="cards_carrito">
@@ -70,6 +101,7 @@ function mostrarCarrito() {
                 </div>
                 <p>${carrito.name}</p>
                 <p>$ ${carrito.precio}</P>
+                <p> x${carrito.cantidad}</p>
             </div>
             <div class="eliminar">
                 <button onClick="eliminar_producto(${carrito.id})" class="btn-eliminar">Eliminar</button>
@@ -84,7 +116,18 @@ function eliminar_producto(id_seleccionado) {
     let eliminar = carrito.filter((e) => e.id != id_seleccionado)
     carrito = eliminar;
     localStorage.setItem("StorageProductos", JSON.stringify(carrito))
-    location.reload();
+
+    Toastify({ //Libreria toastify
+        text: "Producto eliminado del carrito",
+        duration: 1500,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        className: "notificacion-1",
+    }).showToast();
+
+    setTimeout(() => {
+        location.reload()
+    }, 1000)
 }
 console.log(carrito)
 
@@ -96,8 +139,9 @@ let resultado = 0;
 let cuotas = 0;
 
 calcular.onclick = () => {
+
     for (let i = 0; i < carrito.length; i++) {
-        resultado = carrito[i].precio + resultado;
+        resultado = carrito[i].precio * carrito[i].cantidad + resultado;
     }
     console.log(resultado)
     let calculo = document.getElementById("calculo")
@@ -111,22 +155,71 @@ calcular.onclick = () => {
 }
 
 limpiar.onclick = () => {
-    calculo.innerHTML = `$ `
+    location.reload();
 }
-let pagar = document.getElementById("pagar")
 
+let pagar = document.getElementById("pagar")
 pagar.onclick = () => {
     let cuotas = 0;
 
     cuotas = document.getElementById("cuotas").value;
-    contenedor_cuotas =document.getElementById("contenedor_cuotas");
+    contenedor_cuotas = document.getElementById("contenedor_cuotas");
 
     console.log(cuotas)
-    if (cuotas > 0 && cuenta > 0)  {
+    if (cuotas > 0 && cuenta > 0) {
         cu = cuenta / cuotas;
         console.log(cu)
         cu = cu.toFixed(2)
 
         contenedor_cuotas.innerHTML = `<p>Total $${cuenta} en ${cuotas} cuotas de $ ${cu}</p>`
     }
+}
+
+//Filtrar productos
+let todos = document.getElementById("btn-todos")
+let mates = document.getElementById("btn-mates")
+let yerba = document.getElementById("btn-yerba")
+let bombilla = document.getElementById("btn-bombilla")
+
+todos.onclick = () => {
+    filtros("todos");
+}
+mates.onclick = () => {
+    filtros("mate");
+}
+yerba.onclick = () => {
+    filtros("yerba");
+}
+bombilla.onclick = () => {
+    filtros("bombilla");
+}
+
+function filtros(e) {
+    if (e == "todos") {
+        contenedor_productos.innerHTML = ``;
+        cargoProductos();
+    } else {
+        /* console.log(e); */
+        contenedor_productos.innerHTML = ``;
+        productos.forEach(productos => {
+
+            productos.categoria == e ?
+                contenedor_productos.innerHTML += `
+            <div class="card">
+                <div class="foto_producto">
+                    <img class="imagen" src="${productos.imagen}">
+                </div>
+                <div class="card_body">
+                    <p>${productos.name}</p>
+                    <p>$ ${productos.precio}</P>
+                    <div class="flex-row">
+                        <input type="numero" id="cant${productos.id}" placeholder="Cantididad..."> 
+                        <button class="btn-comprar" onClick="seleccionProductos(${productos.id})">Comprar</button>
+                    </div>
+                </div>
+            </div>
+        ` : false
+        });
+    }
+
 }
